backend/src/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.41.0
python-multipart
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings
httpx==0.28.1

# AI/ML
groq==1.0.0
langchain==0.3.15
langchain-groq==0.2.1
duckduckgo-search==6.3.11

# Database
supabase==2.28.0

# Resume Parsing
Pillow
PyMuPDF
pytesseract
